requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-anthropic==0.1.23
langgraph==0.1.19
chromadb==0.5.0
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.0
python-dotenv==1.0.1
openai==1.51.0
tiktoken==0.8.0
numpy==1.26.4
pandas==2.2.3
requests==2.32.3
python-multipart==0.0.12